\name{fourplot}
\alias{fourplot}
\title{The Fourplot}
\usage{
fourplot(x)
}
\arguments{
\item{x}{data to be analyzed}
}
\value{
returns the four plots that is described in one object
}
\description{
Given a dataset it makes four different plots : Sequence Plot, Lag Plot, Histogram and QQ Plot
}
\examples{
fourplot(x)
}
