\name{hint}
\alias{hint}
\title{The Hint function}
\usage{
hint(x = NULL)
}
\description{
The Hint function ist used to get some hints for which parts are wrong in the PPCC part.
}
\arguments{
\item{x}{Describes which particular hint should show up. \code{x} must be a number between 1 and 4. Default is set to \code{NULL}.}
}
\examples{
# No examples for this function.
}
