% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourplot.R

\name{fourplot}
\alias{fourplot}
\title{Four Plot}
\usage{
fourplot(x)
}
\arguments{
\item{x}{A vector of numerical values to be analysed}
\item{xlab}{An optional string vector specifying the x axes}
\item{ylab}{An optional string vector specifying the y axes}
\item{main}{An optional string vector specifying the title of the four plot}
}
\value{
Returns a four plot containing a run sequence plot, a lag plot, an histogram and a normal probability plot to test normal distribution.
}
\description{
fourplot gives you four plots for data analysis.
}
\details{
fourplot plots four functions that can be used for data analysis.
}
\examples{
# Simulate a dataset from a Poisson distribution:

d <- rpois(2000,4)

#Find the fourplot of the function :

fourplot(d)

}
